requirements-base.txt
streamlit==1.36.0
chromadb==0.5.5
sentence-transformers==3.0.1
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
pypdf==4.3.1
safetensors>=0.4.5
